Wisata 2

+++ application/controllers/B_wisata.php
@@ -16,6 +16,9 @@ class B_wisata extends CI_Controller
 
     public function simpan($id = null)
     {
+        // ambil data session user
+        $id_user = $this->session->userdata('id_user');
+
         $rules = $this->Wisata->rules();
         $this->form_validation->set_rules($rules);
 
@@ -29,6 +32,7 @@ class B_wisata extends CI_Controller
                 $foto_wc = fileUpload($_FILES['wc'], 'assets/upload/wisata/');
 
                 $data = array(
+                    'id_user' => $id_user,
                     'nama_wisata' => $_POST['nama_wisata'],
                     'alamat' => $_POST['alamat'],
                     'pusat_informasi' => $_POST['pusat_informasi'],
@@ -44,38 +48,57 @@ class B_wisata extends CI_Controller
                 $simpan = $this->Wisata->simpan($data);
                 echo json_encode(['pesan' => $simpan]);
             } else {
-                $path_parkir = $_POST['path_parkir'];
-                $path_wc = $_POST['path_wc'];
-                $path_mushola = $_POST['path_mushola'];
-                $path_p3k = $_POST['path_p3k'];
+                if ($_POST['p3k'] = 'undefined' || $_POST['mushola'] == 'undefined' || $_POST['tempat_parkir'] == 'undefined' || $_POST['wc'] == 'undefined') {
+                    $data = array(
+                        'id_user' => $id_user,
+                        'nama_wisata' => $_POST['nama_wisata'],
+                        'alamat' => $_POST['alamat'],
+                        'pusat_informasi' => $_POST['pusat_informasi'],
+                        'luas_mushola' => $_POST['luas_mushola'],
+                        'luas_tempat_parkir' => $_POST['luas_tempat_parkir'],
+                        'jumlah_wc' => $_POST['jumlah_wc'],
+                    );
+                } else {
+                    $path_parkir = "assets/upload/wisata/" . $_POST['path_parkir'];
+                    $path_wc = "assets/upload/wisata/" . $_POST['path_wc'];
+                    $path_mushola = "assets/upload/wisata/" . $_POST['path_mushola'];
+                    $path_p3k = "assets/upload/wisata/" . $_POST['path_p3k'];
 
-                if (file_exists($path_parkir, $path_wc, $path_mushola, $path_p3k)) {
-                    unlink($path_parkir);
-                    unlink($path_wc);
-                    unlink($path_mushola);
-                    unlink($path_p3k);
+                    if (file_exists($path_parkir, $path_wc, $path_mushola, $path_p3k)) {
+                        unlink($path_parkir);
+                        unlink($path_wc);
+                        unlink($path_mushola);
+                        unlink($path_p3k);
+                    }
+
+                    $filename_p3k = fileUpload($_FILES['p3k'], 'assets/upload/wisata/');
+                    $filename_mushola = fileUpload($_FILES['mushola'], 'assets/upload/wisata/');
+                    $filename_parkir = fileUpload($_FILES['tempat_parkir'], 'assets/upload/wisata/');
+                    $filename_wc = fileUpload($_FILES['wc'], 'assets/upload/wisata/');
+
+                    $data = array(
+                        'id_user' => $id_user,
+                        'nama_wisata' => $_POST['nama_wisata'],
+                        'alamat' => $_POST['alamat'],
+                        'pusat_informasi' => $_POST['pusat_informasi'],
+                        'p3k' =>  $filename_p3k,
+                        'mushola' =>  $filename_mushola,
+                        'luas_mushola' => $_POST['luas_mushola'],
+                        'tempat_parkir' =>  $filename_parkir,
+                        'luas_tempat_parkir' => $_POST['luas_tempat_parkir'],
+                        'wc' =>  $filename_wc,
+                        'jumlah_wc' => $_POST['jumlah_wc'],
+                    );
                 }
-
-                $filename_p3k = fileUpload($_FILES['p3k'], 'assets/upload/wisata/');
-                $filename_mushola = fileUpload($_FILES['mushola'], 'assets/upload/wisata/');
-                $filename_parkir = fileUpload($_FILES['tempat_parkir'], 'assets/upload/wisata/');
-                $filename_wc = fileUpload($_FILES['wc'], 'assets/upload/wisata/');
-
-                $data = array(
-                    'nama_wisata' => $_POST['nama_wisata'],
-                    'alamat' => $_POST['alamat'],
-                    'pusat_informasi' => $_POST['pusat_informasi'],
-                    'p3k' =>  $filename_p3k,
-                    'mushola' =>  $filename_mushola,
-                    'luas_mushola' => $_POST['luas_mushola'],
-                    'tempat_parkir' =>  $filename_parkir,
-                    'luas_tempat_parkir' => $_POST['luas_tempat_parkir'],
-                    'wc' =>  $filename_wc,
-                    'jumlah_wc' => $_POST['jumlah_wc'],
-                );
                 $edit = $this->Wisata->update($data, $_POST['id']);
                 echo json_encode(['pesan' => $edit]);
             }
         }
+    }
+
+    public function delete($id)
+    {
+        $hapus = $this->Wisata->delete($id);
+        echo json_encode(['pesan' => $hapus]);
     }
 }
